move calculateAnchorPosition to js/lib
