+++ js/lib/calculate-anchor-position.js
